AliaksandrBabtsou/requirements.txt
djangorestframework 
pyyaml
uritemplate
django-cors-headers
psycopg2-binary
requests
python-dateutil
fpdf2